// src/app/lib/generated/models/FantasyRace.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const FantasyRace = {
    Alien: 'alien',
    Android: 'android',
    Barbarian: 'barbarian',
    Cyborg: 'cyborg',
    DarkElf: 'dark_elf',
    DragonQueen: 'dragon_queen',
    Dryad: 'dryad',
    ForestNymph: 'forest_nymph',
    Elf: 'elf',
    ElfPriestess: 'elf_priestess',
    Fairy: 'fairy',
    Ghost: 'ghost',
    Goddess: 'goddess',
    Human: 'human',
    Mage: 'mage',
    Mermaid: 'mermaid',
    Orc: 'orc',
    Succubus: 'succubus',
    Superhero: 'superhero',
    Tentacle: 'tentacle',
    Valkyrie: 'valkyrie',
    Witch: 'witch'
} as const;
export type FantasyRace = typeof FantasyRace[keyof typeof FantasyRace];


export function instanceOfFantasyRace(value: any): boolean {
    for (const key in FantasyRace) {
        if (Object.prototype.hasOwnProperty.call(FantasyRace, key)) {
            if (FantasyRace[key as keyof typeof FantasyRace] === value) {
                return true;
            }
        }
    }
    return false;
}

export function FantasyRaceFromJSON(json: any): FantasyRace {
    return FantasyRaceFromJSONTyped(json, false);
}

export function FantasyRaceFromJSONTyped(json: any, ignoreDiscriminator: boolean): FantasyRace {
    return json as FantasyRace;
}

export function FantasyRaceToJSON(value?: FantasyRace | null): any {
    return value as any;
}

export function FantasyRaceToJSONTyped(value: any, ignoreDiscriminator: boolean): FantasyRace {
    return value as FantasyRace;
}